最新のmasterをマージ

+++ Project/GameSources/ComputeShader.cpp
@@ -9,11 +9,11 @@
 
 namespace basecross {
     bool ComputeShaderUtil::CreateElementBuffer(ID3D11Device* device, D3D11_BUFFER_DESC* desc, BufferContext& context) {
-        desc->ByteWidth = context.GetFullDataSize();
+        desc->ByteWidth = (UINT)context.GetFullDataSize();
         desc->Usage = D3D11_USAGE_DEFAULT;
         desc->BindFlags = D3D11_BIND_UNORDERED_ACCESS | D3D11_BIND_SHADER_RESOURCE;
         desc->MiscFlags = D3D11_RESOURCE_MISC_BUFFER_STRUCTURED;
-        desc->StructureByteStride = context.m_BufferSize;
+        desc->StructureByteStride = (UINT)context.m_BufferSize;
 
         auto result = device->CreateBuffer(desc, nullptr, &context.m_Buffer);
 
@@ -35,7 +35,7 @@ namespace basecross {
         D3D11_UNORDERED_ACCESS_VIEW_DESC accessBufferDesc = {};
         accessBufferDesc.Format = DXGI_FORMAT_UNKNOWN;
         accessBufferDesc.ViewDimension = D3D11_UAV_DIMENSION_BUFFER;
-        accessBufferDesc.Buffer.NumElements = context.m_ArraySize;
+        accessBufferDesc.Buffer.NumElements = (UINT)context.m_ArraySize;
 
         auto result = device->CreateUnorderedAccessView(context.m_Buffer.Get(), &accessBufferDesc, &context.m_UAV);
 
@@ -43,10 +43,10 @@ namespace basecross {
     }
     bool ComputeShaderUtil::CreateReadBackBuffer(ID3D11Device* device, BufferContext& context) {
         D3D11_BUFFER_DESC resultBufferDesc = {};
-        resultBufferDesc.ByteWidth = context.GetFullDataSize();
+        resultBufferDesc.ByteWidth = (UINT)context.GetFullDataSize();
         resultBufferDesc.Usage = D3D11_USAGE_STAGING;
         resultBufferDesc.CPUAccessFlags = D3D11_CPU_ACCESS_READ;
-        resultBufferDesc.StructureByteStride = context.m_BufferSize;
+        resultBufferDesc.StructureByteStride = (UINT)context.m_BufferSize;
 
         auto result = device->CreateBuffer(&resultBufferDesc, nullptr, &context.m_ReadBackBuffer);
